Add settings page dialog placeholder
