+++ README.md
@@ -1,11 +1,9 @@
-# fnordcredit [![Travis badge](https://travis-ci.org/fnordcredit/fnordcredit.svg?branch=master)](https://travis-ci.org/fnordcredit/fnordcredit)
+# Fnordcredit [![Travis badge](https://travis-ci.org/fnordcredit/fnordcredit.svg?branch=master)](https://travis-ci.org/fnordcredit/fnordcredit)
 Open source credit system
 
 Innovative, easy to use credit system for multiple users that comes with an intuitive design: Create an account and charge or discharge your credit.
 
-This is the backend repository, for the frontend look at https://github.com/fnordcredit/frontend.
-
-fnordcredit is written in Javascript/Node.js/react/knex.
+Fnordcredit is written in Typescript/Next.js/tailwindcss/prisma.
 
 # Set up
 To start a local development server do the following:
