+++ tsconfig.json
@@ -25,7 +25,7 @@
         "lib/*"
       ],
       "@components/*": [
-        "components/*"
+        "app/components/*"
       ]
     },
     "plugins": [
